Large refactor of cantons model code
We now differentiate between 3 types of data:
- (Model) parameters: the scalars beta, mu, alpha etc.
- ModelData: data used by the model (population, canton matrix Mij...)
- ValidationData: measured number of cases per canton

In principle, parameters and ModelData may be merged, but we split them
for convenience and performance (see `korali_setup.cpp`).

The data fetch functions are slowly being moved to
`epidemics/data/swiss_cantons.py` and the subfolders therein.
The goal is to get rid of `epidemics/cantons/data`.