SEIIN Interventions unit tests (kind of)
Due to the discontinuity in the interventions (beta and theta), the C++
and Python integrators cannot produce a consistent result. As a
compromise, I've set b1==b2==b3 and theta1==theta2==theta3.

Note that the relative error increases even up to order of 0.01 with
different values of b1, b2 and b3, or theta1, theta2 and theta3.